Move ConnectionList into own file
